Remove util/hexdump
